include:types.h: Correct int32_t to int64_t typo

+++ include/alcp/types.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2021, Advanced Micro Devices. All rights reserved.
+ * Copyright (C) 2019-2022, Advanced Micro Devices. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -44,7 +44,7 @@ typedef enum
 typedef int8_t  Int8;
 typedef int16_t Int16;
 typedef int32_t Int32;
-typedef int32_t Int64;
+typedef int64_t Int64;
 
 typedef uint8_t  Uint8;
 typedef uint16_t Uint16;
